migrate http tests
